+++ Tibby/Tibby/TibbyApp.swift
@@ -12,7 +12,6 @@ import UIKit
 //App Delegate
 class AppDelegate: NSObject, UIApplicationDelegate {
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplication.LaunchOptionsKey : Any]? = nil) -> Bool {
-//        print("JORGE App Did Launch!")
         return true
     }
     
@@ -32,7 +31,6 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     
     func scene(_ scene: UIScene, willConnectTo session: UISceneSession, options connectionOptions: UIScene.ConnectionOptions) {
         guard let _ = (scene as? UIWindowScene) else { return }
-//        print("JORGE SceneDelegate is connected!")
     }
     
     func sceneDidBecomeActive(_ scene: UIScene) {
