+++ Tibby/Tibby/Views/Gatcha/GatchaView.swift
@@ -16,7 +16,7 @@ struct GatchaView: View {
     @State private var snappedItem = 0.0
     @State private var draggingItem = 0.0
     @State private var xOffset: CGFloat = 0
-    @Binding var isBaseOnFocus: Bool
+    @State var isBaseOnFocus: Bool 
     @State private var pressedButton = false
     @State private var backgroundcolor = Color.tibbyBaseWhite
     @State private var disableButton = false
