+++ Tibby/Tibby/Views/OnboardingViews/OnboardingTab.swift
@@ -75,7 +75,7 @@ struct OnboardingTab: View {
                             service.getUser()?.username = name
                         }
                 case .gacha:
-                    GatchaView(isBaseOnFocus: .constant(true), firstTimeHere: $firstTime, currentTibby: $currentTibby)
+                    GatchaView(isBaseOnFocus: true,firstTimeHere: $firstTime, currentTibby: $currentTibby)
                         .onAppear(perform: {
                             if let user = service.getUser() {
                                 service.updateUser(user: user, username: name)
