+++ Tibby/Tibby/Views/PopUp/CustomPopUpView.swift
@@ -111,3 +111,5 @@ enum PopUpActionType {
         }
     }
 }
+
+
