+++ Tibby/Tibby/Views/Store/Components/GachaCardComponent.swift
@@ -20,9 +20,6 @@ struct GachaCardComponent: View {
     private var title: String
     private var color: Color
     private var image: String
-    private var timeTillNextSeries: String {
-        viewModel.countdownUntilNextCollection()
-    }
     
     init(gachaType: GachaType, title: String, color: Color, image: String) {
         self.gachaType = gachaType
@@ -89,46 +86,35 @@ struct GachaCardComponent: View {
         .frame(width: 180, height: 238)
     }
     func startTimer() {
-        updateCountdown()  // Update immediately when the view appears
+        updateCountdown()
         timer = Timer.scheduledTimer(withTimeInterval: 1.0, repeats: true) { _ in
             updateCountdown()
         }
     }
     
-    // Function to stop the timer when it's no longer needed
     func stopTimer() {
         timer?.invalidate()
         timer = nil
     }
     
-    // Function to calculate and update the countdown string
     func updateCountdown() {
         let currentDate = Date()
         let calendar = Calendar.current
-        
-        // Get the current week and year
         let currentWeek = calendar.component(.weekOfYear, from: currentDate)
         let currentYear = calendar.component(.yearForWeekOfYear, from: currentDate)
         
-        // Calculate the start of the next week (e.g., Monday 00:00)
         if let nextWeekStart = calendar.date(from: DateComponents(weekOfYear: currentWeek + 1, yearForWeekOfYear: currentYear)) {
-            // Calculate the time interval between the current date and the start of the next week
             let timeInterval = nextWeekStart.timeIntervalSince(currentDate)
             
-            // If timeInterval is less than or equal to zero, stop the timer
             if timeInterval <= 0 {
                 stopTimer()
-                countdownString = "00:00:00"  // Countdown finished
+                countdownString = "00:00:00"
                 return
             }
-            
-            // Convert the time interval into hours, minutes, and seconds
             let totalSeconds = Int(timeInterval)
             let hours = totalSeconds / 3600
             let minutes = (totalSeconds % 3600) / 60
             let seconds = totalSeconds % 60
-            
-            // Format the output as "HH:mm:ss"
             countdownString = String(format: "%02d:%02d:%02d", hours, minutes, seconds)
         }
     }
